Split unit tests which are 100% internal/local/offline from external assets needed via integration test suite
