+++ src/test/java/org/openqa/selenium/htmlunit/HtmlUnitDriverIT.java
@@ -1,0 +1,51 @@
+// Licensed to the Software Freedom Conservancy (SFC) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The SFC licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.openqa.selenium.htmlunit;
+
+import org.htmlunit.BrowserVersion;
+import org.htmlunit.WebClient;
+import org.junit.Test;
+import org.openqa.selenium.Platform;
+import org.openqa.selenium.Proxy;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.htmlunit.options.HtmlUnitDriverOptions;
+import org.openqa.selenium.remote.Browser;
+import org.openqa.selenium.remote.CapabilityType;
+import org.openqa.selenium.remote.DesiredCapabilities;
+
+import static org.junit.Assert.*;
+
+/**
+ * General tests for the HtmlUnitDriver. (External Domain Tests)
+ *
+ * @author Ronald Brill
+ * @author Martin Bartoš
+ * @author Scott Babcock
+ */
+public class HtmlUnitDriverIT {
+
+    @Test
+    public void resetWebClient() {
+        final HtmlUnitDriver webDriver = new HtmlUnitDriver();
+
+        webDriver.get("https://www.htmlunit.org");
+        webDriver.getWebClient().reset();
+
+        webDriver.get("https://www.htmlunit.org");
+    }
+}
