+++ src/test/java/org/openqa/selenium/htmlunit/HtmlUnitDriverTest.java
@@ -542,14 +542,4 @@ public class HtmlUnitDriverTest {
             }
         };
     }
-
-    @Test
-    public void resetWebClient() {
-        final HtmlUnitDriver webDriver = new HtmlUnitDriver();
-
-        webDriver.get("https://www.htmlunit.org");
-        webDriver.getWebClient().reset();
-
-        webDriver.get("https://www.htmlunit.org");
-    }
 }
